requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
APScheduler==3.10.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.3
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
Flask==3.0.0
Flask-Cors==4.0.0
frozenlist==1.4.0
google-auth==2.23.3
google-auth-oauthlib==1.1.0
greenlet==3.0.0
gspread==5.11.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.26.1
oauthlib==3.2.2
openai==0.28.1
packaging==23.2
pandas==2.1.1
playwright==1.39.0
plotly==5.17.0
pyahocorasick==2.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyee==11.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
slack-sdk==3.23.0
soupsieve==2.5
SQLAlchemy==2.0.22
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
Werkzeug==3.0.0
yarl==1.9.2
